+++ textx/cli.py
@@ -3,7 +3,7 @@ import sys
 import click
 import pkg_resources
 from textx import metamodel_from_file, TextXError
-from textx.export import metamodel_export, model_export
+from textx.export import metamodel_export, model_export, PlantUmlRenderer
 
 
 @click.group()
@@ -19,6 +19,11 @@ model_arg = click.argument('model_file', type=click.Path(), required=False)
 ignore_case_arg = click.option('--ignore-case/', '-i/', default=False,
                                is_flag=True,
                                help="case-insensitive parsing.")
+format_arg = click.option('--output-format', '-f',
+                          type=click.Choice(['dot', 'plantuml']),
+                          default='dot',
+                          help="select the output format (plantuml not "
+                               "available for model files, yet).")
 
 
 @textx.command()
@@ -38,20 +43,30 @@ def check(ctx, meta_model_file, model_file, ignore_case):
 @meta_model_arg
 @model_arg
 @ignore_case_arg
+@format_arg
 @click.pass_context
-def visualize(ctx, meta_model_file, model_file, ignore_case):
+def visualize(ctx, meta_model_file, model_file, ignore_case, output_format):
     """
     Generate .dot file(s) from meta-model and optionally model.
     """
     debug = ctx.obj['debug']
     meta_model, model = check_model(meta_model_file,
                                     model_file, debug, ignore_case)
-    dot_file = "{}.dot".format(meta_model_file)
-    click.echo("Generating '{}' file for meta-model.".format(dot_file))
-    click.echo("To convert to png run 'dot -Tpng -O {}'".format(dot_file))
-    metamodel_export(meta_model, dot_file)
+    if output_format == 'plantuml':
+        pu_file = "{}.pu".format(meta_model_file)
+        click.echo("Generating '{}' file for meta-model.".format(pu_file))
+        click.echo("To convert to png run 'plantuml {}'".format(pu_file))
+        click.echo("To convert to svg run 'plantuml -tsvg {}'".format(pu_file))
+        metamodel_export(meta_model, pu_file, PlantUmlRenderer())
+    else:
+        dot_file = "{}.dot".format(meta_model_file)
+        click.echo("Generating '{}' file for meta-model.".format(dot_file))
+        click.echo("To convert to png run 'dot -Tpng -O {}'".format(dot_file))
+        metamodel_export(meta_model, dot_file)
 
     if model_file:
+        if output_format == 'plantuml':
+            raise Exception("plantuml is not supported for model files, yet.")
         dot_file = "{}.dot".format(model_file)
         click.echo("Generating '{}' file for model.".format(model_file))
         click.echo(