+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # textX changelog
 
 * Next Release 1.?.?
-
    - https://github.com/textX/textX/pull/168
      - Adding examples and documentation related to scope providers
        (related to model modification through scope providers).
+   - https://github.com/textX/textX/pull/165
+     - Added metamodel export feature for PlantUML.
    - https://github.com/textX/textX/pull/134
      - Fixing bug when using a sequence of matches and rule reference in
        an abstract rule choice alternative.